+++ intelmq/lib/bot.py
@@ -337,7 +337,12 @@ class Bot(object):
             self.__destination_pipeline = None
             self.logger.debug("Disconnected from destination pipeline.")
 
-    def send_message(self, *messages):
+    def send_message(self, *messages, auto_add=None):
+        """
+        Parameters:
+            messages: Instances of intelmq.lib.message.Message class
+            auto_add: ignored
+        """
         for message in messages:
             if not message:
                 self.logger.warning("Ignoring empty message at sending. Possible bug in bot.")
@@ -710,9 +715,15 @@ class CollectorBot(Bot):
         report.add("feed.accuracy", self.parameters.accuracy)
         return report
 
-    def send_message(self, *messages):
+    def send_message(self, *messages, auto_add=True):
+        """"
+        Parameters:
+            messages: Instances of intelmq.lib.message.Message class
+            auto_add: Add some default report fields form parameters
+        """
         messages = filter(self.__filter_empty_report, messages)
-        messages = map(self.__add_report_fields, messages)
+        if auto_add:
+            messages = map(self.__add_report_fields, messages)
         super(CollectorBot, self).send_message(*messages)
 
     def new_report(self):