+++ CHANGELOG.md
@@ -39,6 +39,7 @@ CHANGELOG
 - `intelmqdump` automatically converts messages from events to reports if the queue the message is being restored to is the source queue of a parser (#1225).
 - `intelmqdump` is now capable to read messages in dumps that are dictionaries as opposed to serialized dicts as strings and does not convert them in the show command (#1256).
 - `intelmqdump` truncated messages are no longer used/saved to the file after being shown (#1255).
+- `intelmqctl run [bot-id] mesage send` can now send report messages (#1077).
 
 ### Contrib